requirements.txt
# Core dependencies
python-telegram-bot
pydantic
pydantic-settings
python-dotenv

# Database
aiosqlite

# Translation
openai
deep-translator

# Task scheduling
APScheduler
nest_asyncio

# Utilities
tiktoken
numpy

# Type hints
typing-extensions